BREAKING CHANGE: new data and error handling

+++ src/routes/(main)/settings/+page.server.ts
@@ -1,33 +1,60 @@
 import { fail, type Actions, redirect } from '@sveltejs/kit';
 import { env } from '$env/dynamic/private';
+import { fileURLToPath } from 'url';
 import type { PageServerLoad } from './$types';
+import type { MyUser } from '$lib/types';
+
+const __filename = fileURLToPath(import.meta.url);
+const __route = __filename.slice(__filename.indexOf('src'));
 
 export const load: PageServerLoad = async ({ fetch, cookies }) => {
   const base_api_url: string = env.API_URL;
   const url = `${base_api_url}/users/myUser`;
-  const access_token = cookies.get('access_token');
-  
-  if (access_token) {
-    const options = {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${access_token}`,
-        'Content-Type': 'application/json'
+  let fetchedMyUser = { data: null, status: 500 };
+  let myUser_response: Promise<{ data: MyUser | null, status: number }> = Promise.resolve(fetchedMyUser);
+
+  if (!base_api_url) {
+    console.error(`Error: Error en [${__route}].\n\t- No se encontro la url de la api en el entorno`);
+    return {
+      streamed: {
+        myUser: myUser_response,
       }
-    };
-    const response = await fetch(url, options);
-    
-    if (response.ok) {
-      return {
-        streamed: {
-          user: response.json()
-        }
-      };
-    } else if (response.status === 401) throw redirect(303, "/login");
-    
-  } else {
-    throw redirect(303, "/login")
+    }
   }
+
+  const access_token = cookies.get('access_token');
+  if (!access_token) {
+    console.error(`Error: Error en [${__route}].\n\t- El usuario no ha iniciado sesion`);
+    throw redirect(303, "/login");
+  }
+
+  const options = {
+    method: 'GET',
+    headers: {
+      Authorization: `Bearer ${access_token}`,
+      'Content-Type': 'application/json'
+    }
+  };
+
+  myUser_response = fetch(url, options)
+    .then(async (response) => {
+      fetchedMyUser.status = response.status;
+      if (response.ok) {
+        const data = await response.json();
+        fetchedMyUser.data = data;
+      }
+      return fetchedMyUser;
+    }).catch((error) => {
+      console.error(`Error: Error en [${__route}].\n\t- Error al intentar obtener "myUser"\n\t- ${error}`)
+      return fetchedMyUser;
+    });
+
+  return {
+    streamed: {
+      myUser: myUser_response,
+    }
+  }
+
 };
 
 export const actions: Actions = {
@@ -38,9 +65,16 @@ export const actions: Actions = {
     const update_url = `${base_api_url}/users/update`;
     const access_token = cookies.get('access_token');
 
+    if (!base_api_url) return fail(500, { serverFail: true });
+    if (!access_token) {
+      console.error(`Error: Error en [${__route}].\n\t- El usuario no ha iniciado sesion`);
+      throw redirect(303, "/login");
+    }
+
     const body = {
-      ...(description ? { description } : {description: ''}),
+      ...(description ? { description } : { description: '' }),
     };
+
     const options = {
       method: 'PUT',
       headers: {
@@ -49,7 +83,20 @@ export const actions: Actions = {
       },
       body: JSON.stringify(body)
     };
-    const response = await fetch(update_url, options);
+
+    let response: Response;
+
+    try {
+      response = await fetch(update_url, options);
+    } catch (error) {
+      console.error(`Error: Error en [${__route}].\n\t- Error al intentar "editDescription"\n\t- ${error}`)
+      return fail(500, { serverFail: true });
+    }
+
+    if (response.ok) return;
+    else if (response.status === 401) throw redirect(303, "/login")
+    else return fail(response.status, { serverFail: true });
+
   },
   editWebSite: async ({ request, fetch, cookies }) => {
     const form = await request.formData();
@@ -57,12 +104,17 @@ export const actions: Actions = {
     const base_api_url: string = env.API_URL;
     const update_url = `${base_api_url}/users/update`;
     const access_token = cookies.get('access_token');
-    console.log(personal_url)
+
+    if (!base_api_url) return fail(500, { serverFail: true });
+    if (!access_token) {
+      console.error(`Error: Error en [${__route}].\n\t- El usuario no ha iniciado sesion`);
+      throw redirect(303, "/login");
+    }
 
     const body = {
-      ...(personal_url ? { personal_url } : {personal_url: ''}),
+      ...(personal_url ? { personal_url } : { personal_url: '' }),
     };
-    console.log(body)
+
     const options = {
       method: 'PUT',
       headers: {
@@ -71,7 +123,20 @@ export const actions: Actions = {
       },
       body: JSON.stringify(body)
     };
-    const response = await fetch(update_url, options);
+
+    let response: Response;
+
+    try {
+      response = await fetch(update_url, options);
+    } catch (error) {
+      console.error(`Error: Error en [${__route}].\n\t- Error al intentar "editWebSite"\n\t- ${error}`)
+      return fail(500, { serverFail: true });
+    }
+
+    if (response.ok) return;
+    else if (response.status === 401) throw redirect(303, "/login")
+    else return fail(response.status, { serverFail: true });
+
   },
   editLocation: async ({ request, fetch, cookies }) => {
     const form = await request.formData();
@@ -80,10 +145,16 @@ export const actions: Actions = {
     const update_url = `${base_api_url}/users/update`;
     const access_token = cookies.get('access_token');
 
+    if (!base_api_url) return fail(500, { serverFail: true });
+    if (!access_token) {
+      console.error(`Error: Error en [${__route}].\n\t- El usuario no ha iniciado sesion`);
+      throw redirect(303, "/login");
+    }
+
     const body = {
-      ...(location ? { location } : {location: ''}),
+      ...(location ? { location } : { location: '' }),
     };
-    console.log(body)
+
     const options = {
       method: 'PUT',
       headers: {
@@ -92,20 +163,37 @@ export const actions: Actions = {
       },
       body: JSON.stringify(body)
     };
-    const response = await fetch(update_url, options);
+    let response: Response;
+
+    try {
+      response = await fetch(update_url, options);
+    } catch (error) {
+      console.error(`Error: Error en [${__route}].\n\t- Error al intentar "editLocation"\n\t- ${error}`)
+      return fail(500, { serverFail: true });
+    }
+
+    if (response.ok) return;
+    else if (response.status === 401) throw redirect(303, "/login")
+    else return fail(response.status, { serverFail: true });
+
   },
   editBirthday: async ({ request, fetch, cookies }) => {
     const form = await request.formData();
     const birthday = form.get('birthday');
-    console.log(birthday)
     const base_api_url: string = env.API_URL;
     const update_url = `${base_api_url}/users/update`;
     const access_token = cookies.get('access_token');
 
+    if (!base_api_url) return fail(500, { serverFail: true });
+    if (!access_token) {
+      console.error(`Error: Error en [${__route}].\n\t- El usuario no ha iniciado sesion`);
+      throw redirect(303, "/login");
+    }
+
     const body = {
-      ...(birthday ? { birthday } : {birthday: ''}),
+      ...(birthday ? { birthday } : { birthday: '' }),
     };
-    console.log(body)
+
     const options = {
       method: 'PUT',
       headers: {
@@ -114,12 +202,26 @@ export const actions: Actions = {
       },
       body: JSON.stringify(body)
     };
-    const response = await fetch(update_url, options);
+    let response: Response;
+
+    try {
+      response = await fetch(update_url, options);
+    } catch (error) {
+      console.error(`Error: Error en [${__route}].\n\t- Error al intentar "editBirthday"\n\t- ${error}`)
+      return fail(500, { serverFail: true });
+    }
+
+    if (response.ok) return;
+    else if (response.status === 401) throw redirect(303, "/login");
+    else return fail(response.status, { serverFail: true });
   },
   changeUsername: async ({ request, fetch, cookies }) => {
     const form = await request.formData();
     const username = form.get('username');
     const password = form.get('password');
+    const base_api_url: string = env.API_URL;
+    const url = `${base_api_url}/users/username`;
+    const access_token = cookies.get('access_token');
 
     if (!username) {
       return fail(400, { usernameMissing: true });
@@ -129,53 +231,82 @@ export const actions: Actions = {
       return fail(400, { passwordMissing: true });
     }
 
-    const base_api_url: string = env.API_URL;
-    const url = `${base_api_url}/users/username`;
-    const access_token = cookies.get('access_token');
-    if (access_token) {
-      const body = {
-        username,
-        password
-      };
+    if (!base_api_url) return fail(500, { serverFail: true });
+
+    if (!access_token) {
+      console.error(`Error: Error en [${__route}].\n\t- El usuario no ha iniciado sesion`);
+      throw redirect(303, "/login");
+    }
+
+    const body = {
+      username,
+      password
+    };
+
+    const options = {
+      method: 'PUT',
+      headers: {
+        Authorization: `Bearer ${access_token}`,
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(body)
+    };
+
+    let response: Response;
+
+    try {
+      response = await fetch(url, options);
+    } catch (error) {
+      console.error(`Error: Error en [${__route}].\n\t- Error al intentar "changeUsername"\n\t- ${error}`)
+      return fail(500, { serverFail: true });
+    }
+
+    if (response.ok) {
+
+      const body = new URLSearchParams();
+      body.append('username', username.toString());
+      body.append('password', password.toString());
       const options = {
-        method: 'PUT',
+        method: 'POST',
         headers: {
-          Authorization: `Bearer ${access_token}`,
-          'Content-Type': 'application/json'
+          'Content-Type': 'application/x-www-form-urlencoded'
         },
-        body: JSON.stringify(body)
+        body: body.toString()
       };
-      const response = await fetch(url, options);
-      if (response.status === 400) {
-        return fail(400, { usernameExists: true });
-      } else if (response.status === 404) {
-        return fail(404, { usernameNotFound: true });
-      } else if (response.status === 401) {
-        return fail(401, { wrongPassword: true });
-      } else if (response.ok) {
-        const body = new URLSearchParams();
-        body.append('username', username.toString());
-        body.append('password', password.toString());
-        const options = {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/x-www-form-urlencoded'
-          },
-          body: body.toString()
-        };
-        const url = `${base_api_url}/users/signin`;
-        const response = await fetch(url, options);
-        if (response.ok) {
-          const responseJSON = await response.json();
-          cookies.set('access_token', responseJSON.access_token, {
-            path: '/'
-          });
-          cookies.set('username', responseJSON.username, {
-            path: '/'
-          });
-          throw redirect(303, '/settings');
-        }
+
+      const signin_url = `${base_api_url}/users/signin`;
+      let signin_response: Response;
+
+      try {
+        signin_response = await fetch(signin_url, options);
+      } catch (error) {
+        console.error(`Error: Error en [${__route}].\n\t- Error al intentar "signin"\n\t- ${error}`)
+        return fail(500, { serverFail: true });
       }
+
+      if (signin_response.ok) {
+        const responseJSON = await signin_response.json();
+        cookies.set('access_token', responseJSON.access_token, {
+          path: '/'
+        });
+        cookies.set('username', responseJSON.username, {
+          path: '/'
+        });
+        return;
+      } else if (signin_response.status == 401) {
+        throw redirect(303, "/login")
+      } else {
+        return fail(response.status, { serverFail: true });
+      }
+
+    } else if (response.status === 400) {
+      return fail(400, { usernameExists: true });
+    } else if (response.status === 404) {
+      return fail(404, { usernameNotFound: true });
+    } else if (response.status === 401) {
+      return fail(401, { wrongPassword: true });
+    } else {
+      return fail(response.status, { serverFail: true });
     }
   },
   changePassword: async ({ request, fetch, cookies }) => {
@@ -183,6 +314,16 @@ export const actions: Actions = {
     const password = form.get('password');
     const newPassword = form.get('new_password');
     const confirm = form.get('confirm');
+    const base_api_url: string = env.API_URL;
+    const url = `${base_api_url}/users/passsword`;
+    const access_token = cookies.get('access_token');
+
+    if (!base_api_url) return fail(500, { serverFail: true });
+
+    if (!access_token) {
+      console.error(`Error: Error en [${__route}].\n\t- El usuario no ha iniciado sesion`);
+      throw redirect(303, "/login");
+    }
 
     if (!password) {
       return fail(400, { passwordMissing: true });
@@ -197,32 +338,48 @@ export const actions: Actions = {
       return fail(400, { passwordConfirm: true });
     }
 
-    const base_api_url: string = env.API_URL;
-    const url = `${base_api_url}/users/passsword`;
-    const access_token = cookies.get('access_token');
-    if (access_token) {
-      const body = {
-        password: password,
-        new_password: newPassword
-      };
-      const options = {
-        method: 'PUT',
-        headers: {
-          Authorization: `Bearer ${access_token}`,
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(body)
-      };
-      const response = await fetch(url, options);
-      if (response.ok) {
-        throw redirect(303, '/settings');
-      }
+    const body = {
+      password: password,
+      new_password: newPassword
+    };
+    const options = {
+      method: 'PUT',
+      headers: {
+        Authorization: `Bearer ${access_token}`,
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(body)
+    };
+
+    let response: Response;
+
+    try {
+      response = await fetch(url, options);
+    } catch (error) {
+      console.error(`Error: Error en [${__route}].\n\t- Error al intentar "changePassword"\n\t- ${error}`)
+      return fail(500, { serverFail: true });
+    }
+
+    if (response.ok) {
+      return;
+    } else if (response.status === 401) {
+      throw redirect(303, "/login")
+    } else {
+      return fail(response.status, { serverFail: true });
     }
   },
   deleteAccount: async ({ fetch, cookies }) => {
     const base_api_url: string = env.API_URL;
     const url = `${base_api_url}/users/`;
     const access_token = cookies.get('access_token');
+
+    if (!base_api_url) return fail(500, { serverFail: true });
+
+    if (!access_token) {
+      console.error(`Error: Error en [${__route}].\n\t- El usuario no ha iniciado sesion`);
+      throw redirect(303, "/login");
+    }
+
     const options = {
       method: 'DELETE',
       headers: {
@@ -230,7 +387,16 @@ export const actions: Actions = {
         'Content-Type': 'application/json'
       }
     };
-    const response = await fetch(url, options);
+
+    let response: Response;
+
+    try {
+      response = await fetch(url, options);
+    } catch (error) {
+      console.error(`Error: Error en [${__route}].\n\t- Error al intentar "deleteAccount"\n\t- ${error}`)
+      return fail(500, { serverFail: true });
+    }
+
     if (response.ok) {
       cookies.delete('access_token', { path: '/' });
       cookies.delete('username', { path: '/' });
@@ -238,6 +404,9 @@ export const actions: Actions = {
       cookies.delete('user_id', { path: '/' });
       cookies.delete('refresh_token', { path: '/' });
       throw redirect(303, '/home');
-    }
+    } else if (response.status == 401) {
+      throw redirect(303, "/login");
+    } else return fail(response.status, { serverFail: true });
+
   }
 };
